basic of code reuse - extract base class for both F16 anf FA18 - refactoring of aircraft handling
